data 모듈 클래스 이동 및 viewModel 이동
